+++ packages/core/src/define-api/fixtures/test-plugin/test.api.ts
@@ -1,9 +1,0 @@
-import { defineApi } from '@faasjs/core'
-
-export default defineApi({
-  async handler(data) {
-    return {
-      loaded: Boolean((data.context as Record<string, any>).fileUrlFixturePluginLoaded),
-    }
-  },
-})